Use user-specified font instead of hardcoded list
Replace list with "system fonts" from a handful of operating systems with
the default browser font, which is much easier for the user to change.

For some users, if they are using an OS represented in the list, this might
be the same font, but for those who have explicitly specified a different
font, Mastodon will now use that instead.